# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus AI Platform - AI Recruitment & CFO SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Auth session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Unauthorized
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all platform users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformUser"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a new platform user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformUser"
  /admin/users/{id}:
    get:
      operationId: getUser
      tags: [admin]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformUser"
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update user role or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformUser"
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Deactivate user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/organization:
    get:
      operationId: getOrganization
      tags: [admin]
      summary: Get organization settings
      responses:
        "200":
          description: Organization settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
    patch:
      operationId: updateOrganization
      tags: [admin]
      summary: Update organization settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /admin/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [admin]
      summary: Get platform audit logs
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RegisterInput:
      type: object
      required: [email, password, full_name, organization_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        full_name:
          type: string
        organization_name:
          type: string
    AuthSession:
      type: object
      required: [access_token, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, email, full_name, role]
      properties:
        id:
          type: string
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, recruiter, hr_manager, cfo, manager, viewer]
        avatar_url:
          type: ["string", "null"]
        organization_id:
          type: ["string", "null"]
        organization_name:
          type: ["string", "null"]
        created_at:
          type: string

    # DASHBOARD
...
    PlatformUser:
      type: object
      required: [id, email, full_name, role, status, created_at]
      properties:
        id:
          type: string
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, recruiter, hr_manager, cfo, manager, viewer]
        status:
          type: string
          enum: [active, inactive, pending]
        last_login:
          type: ["string", "null"]
        created_at:
          type: string
    UserInput:
      type: object
      required: [email, full_name, role]
      properties:
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, recruiter, hr_manager, cfo, manager, viewer]
    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
        role:
          type: string
        status:
          type: string
    Organization:
      type: object
      required: [id, name, industry, size]
      properties:
        id:
          type: string
        name:
          type: string
        industry:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        logo_url:
          type: ["string", "null"]
        headquarters:
          type: ["string", "null"]
        founded_year:
          type: ["integer", "null"]
        fiscal_year_start:
          type: ["string", "null"]
        currency:
          type: string
    OrganizationUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        size:
          type: string
        website:
          type: string
        logo_url:
          type: string
        headquarters:
          type: string
        founded_year:
          type: integer
        fiscal_year_start:
          type: string
        currency:
          type: string
    AuditLog:
      type: object
      required: [id, action, entity_type, user_id, user_name, created_at]
      properties:
        id:
          type: string
        action:
          type: string
        entity_type:
          type: string
        entity_id:
          type: ["string", "null"]
        user_id:
          type: string
        user_name:
          type: string
        ip_address:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        created_at:
          type: string