# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus AI Platform - AI Recruitment & CFO SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfile } from './userProfile';

export interface AuthSession {
  access_token: string;
  /** @nullable */
  refresh_token?: string | null;
  user: UserProfile;
}