# lib/api-zod/src/generated/types/platformUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexus AI Platform - AI Recruitment & CFO SaaS
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformUserRole } from './platformUserRole';
import type { PlatformUserStatus } from './platformUserStatus';

export interface PlatformUser {
  id: string;
  email: string;
  full_name: string;
  role: PlatformUserRole;
  status: PlatformUserStatus;
  /** @nullable */
  last_login?: string | null;
  created_at: string;
}